refactor: past_to_itir_wf -> past_to_itir
