+++ src/gt4py/next/ffront/past_to_itir.py
@@ -14,14 +14,93 @@
 
 from __future__ import annotations
 
-from typing import Optional, cast
+import dataclasses
+from typing import Any, Optional, cast
+
+import devtools
+import factory
 
 from gt4py.eve import NodeTranslator, concepts, traits
-from gt4py.next.common import Dimension, DimensionKind, GridType
-from gt4py.next.ffront import lowering_utils, program_ast as past, type_specifications as ts_ffront
+from gt4py.next import common, config
+from gt4py.next.ffront import (
+    fbuiltins,
+    gtcallable,
+    lowering_utils,
+    program_ast as past,
+    transform_utils,
+    type_specifications as ts_ffront,
+)
 from gt4py.next.iterator import ir as itir
 from gt4py.next.iterator.ir_utils import ir_makers as im
+from gt4py.next.otf import stages, workflow
 from gt4py.next.type_system import type_info, type_specifications as ts
+
+
+@dataclasses.dataclass(frozen=True)
+class PastToItir(workflow.ChainableWorkflowMixin):
+    def __call__(self, inp: stages.PastClosure) -> stages.ProgramCall:
+        all_closure_vars = transform_utils._get_closure_vars_recursively(inp.closure_vars)
+        offsets_and_dimensions = transform_utils._filter_closure_vars_by_type(
+            all_closure_vars, fbuiltins.FieldOffset, common.Dimension
+        )
+        grid_type = transform_utils._deduce_grid_type(
+            inp.grid_type, offsets_and_dimensions.values()
+        )
+
+        gt_callables = transform_utils._filter_closure_vars_by_type(
+            all_closure_vars, gtcallable.GTCallable
+        ).values()
+        lowered_funcs = [gt_callable.__gt_itir__() for gt_callable in gt_callables]
+
+        itir_program = ProgramLowering.apply(
+            inp.past_node, function_definitions=lowered_funcs, grid_type=grid_type
+        )
+
+        if config.DEBUG or "debug" in inp.kwargs:
+            devtools.debug(itir_program)
+
+        return stages.ProgramCall(
+            itir_program,
+            inp.args,
+            inp.kwargs | {"column_axis": _column_axis(all_closure_vars)},
+        )
+
+
+class PastToItirFactory(factory.Factory):
+    class Meta:
+        model = PastToItir
+
+
+def _column_axis(all_closure_vars: dict[str, Any]) -> common.Dimension:
+    # construct mapping from column axis to scan operators defined on
+    #  that dimension. only one column axis is allowed, but we can use
+    #  this mapping to provide good error messages.
+    scanops_per_axis: dict[common.Dimension, str] = {}
+    for name, gt_callable in transform_utils._filter_closure_vars_by_type(
+        all_closure_vars, gtcallable.GTCallable
+    ).items():
+        if isinstance(
+            (type_ := gt_callable.__gt_type__()),
+            ts_ffront.ScanOperatorType,
+        ):
+            scanops_per_axis.setdefault(type_.axis, []).append(name)
+
+    if len(scanops_per_axis.values()) == 0:
+        return None
+
+    if len(scanops_per_axis.values()) != 1:
+        scanops_per_axis_strs = [
+            f"- {dim.value}: {', '.join(scanops)}" for dim, scanops in scanops_per_axis.items()
+        ]
+
+        raise TypeError(
+            "Only 'ScanOperator's defined on the same axis "
+            + "can be used in a 'Program', found:\n"
+            + "\n".join(scanops_per_axis_strs)
+            + "."
+        )
+
+    return iter(scanops_per_axis.keys()).__next__()
 
 
 def _size_arg_from_field(field_name: str, dim: int) -> str:
@@ -51,10 +130,10 @@ class ProgramLowering(
     --------
     >>> from gt4py.next.ffront.func_to_past import ProgramParser
     >>> from gt4py.next.iterator import ir
-    >>> from gt4py.next import Dimension, Field
+    >>> from gt4py.next import common.Dimension, Field
     >>>
     >>> float64 = float
-    >>> IDim = Dimension("IDim")
+    >>> IDim = common.Dimension("IDim")
     >>>
     >>> def fieldop(inp: Field[[IDim], "float64"]) -> Field[[IDim], "float64"]: ...
     >>> def program(inp: Field[[IDim], "float64"], out: Field[[IDim], "float64"]):
@@ -67,7 +146,7 @@ class ProgramLowering(
     ...     expr=ir.FunCall(fun=ir.SymRef(id="deref"), pos_only_args=[ir.SymRef(id="inp")]),
     ... )  # doctest: +SKIP
     >>> lowered = ProgramLowering.apply(
-    ...     parsed, [fieldop_def], grid_type=GridType.CARTESIAN
+    ...     parsed, [fieldop_def], grid_type=common.GridType.CARTESIAN
     ... )  # doctest: +SKIP
     >>> type(lowered)  # doctest: +SKIP
     <class 'gt4py.next.iterator.ir.FencilDefinition'>
@@ -85,7 +164,7 @@ class ProgramLowering(
         cls,
         node: past.Program,
         function_definitions: list[itir.FunctionDefinition],
-        grid_type: GridType,
+        grid_type: common.GridType,
     ) -> itir.FencilDefinition:
         return cls(grid_type=grid_type).visit(node, function_definitions=function_definitions)
 
@@ -97,7 +176,7 @@ class ProgramLowering(
         size_params = []
         for param in node.params:
             if type_info.is_type_or_tuple_of_type(param.type, ts.FieldType):
-                fields_dims: list[list[Dimension]] = (
+                fields_dims: list[list[common.Dimension]] = (
                     type_info.primitive_constituents(param.type).getattr("dims").to_list()
                 )
                 assert all(field_dims == fields_dims[0] for field_dims in fields_dims)
@@ -263,8 +342,8 @@ class ProgramLowering(
                     slices[dim_i].upper if slices else None, dim_size, dim_size
                 )
 
-            if dim.kind == DimensionKind.LOCAL:
-                raise ValueError(f"Dimension '{dim.value}' must not be local.")
+            if dim.kind == common.DimensionKind.LOCAL:
+                raise ValueError(f"common.Dimension '{dim.value}' must not be local.")
             domain_args.append(
                 itir.FunCall(
                     fun=itir.SymRef(id="named_range"),
@@ -273,14 +352,14 @@ class ProgramLowering(
             )
             domain_args_kind.append(dim.kind)
 
-        if self.grid_type == GridType.CARTESIAN:
+        if self.grid_type == common.GridType.CARTESIAN:
             domain_builtin = "cartesian_domain"
-        elif self.grid_type == GridType.UNSTRUCTURED:
+        elif self.grid_type == common.GridType.UNSTRUCTURED:
             domain_builtin = "unstructured_domain"
             # for no good reason, the domain arguments for unstructured need to be in order (horizontal, vertical)
-            if domain_args_kind[0] == DimensionKind.VERTICAL:
+            if domain_args_kind[0] == common.DimensionKind.VERTICAL:
                 assert len(domain_args) == 2
-                assert domain_args_kind[1] == DimensionKind.HORIZONTAL
+                assert domain_args_kind[1] == common.DimensionKind.HORIZONTAL
                 domain_args[0], domain_args[1] = domain_args[1], domain_args[0]
         else:
             raise AssertionError()
@@ -294,14 +373,14 @@ class ProgramLowering(
     def _construct_itir_initialized_domain_arg(
         self,
         dim_i: int,
-        dim: Dimension,
+        dim: common.Dimension,
         node_domain: past.Dict,
     ) -> list[itir.FunCall]:
         assert len(node_domain.values_[dim_i].elts) == 2
         keys_dims_types = cast(ts.DimensionType, node_domain.keys_[dim_i].type).dim
         if keys_dims_types != dim:
             raise ValueError(
-                "Dimensions in out field and field domain are not equivalent:"
+                "common.Dimensions in out field and field domain are not equivalent:"
                 f"expected '{dim}', got '{keys_dims_types}'."
             )
